+++ src/darwin/index.ts
@@ -46,6 +46,7 @@ export { software } from './software';
 export { system } from './system';
 export { thunderbolt } from './thunderbolt';
 export { time } from '../time';
+export { serialPorts } from './serial-ports';
 export { usb } from './usb';
 export { users } from './users';
 export { uuid } from './uuid';